add auth, backend tests, add functionality to add delete and re order, improve css/layout, add search ability, add sort ability on front end

+++ database/migrations/2020_10_06_163947_create_books_table.php
@@ -16,7 +16,7 @@ class CreateBooksTable extends Migration
         Schema::create('books', function (Blueprint $table) {
             $table->string('api_id');
             $table->foreignId('user_id');
-            $table->smallInteger('rank');
+            $table->integer('rank');
             $table->timestamps();
             $table->primary(['api_id', 'user_id']);
         });
